+++ changelog.md
@@ -17,10 +17,15 @@
     - Add models and controllers for stocksBought and users
     - Add static functions to models
     - Add aysnc functions to controllers.
+    - Add auth controller, to allow for auth and auth.
+    - Made routes for the get,post,delete etc...
+    - Updated server.js to have app.use for the users, stocksBought and auth js files in our controllers folder.
 
 ## Database
 
     - Create DB Schema
+    - Creat db for selling and buying stocks.
+    - Created a db for users.
 
 ## Misc